Update to dist builds
