+++ support/grunt/options/copy.ts
@@ -26,6 +26,11 @@ export = function(grunt: IGrunt) {
 					if (srcpath === 'src/examples/index.html') {
 						return content.replace('"../../../node_modules/', '"../node_modules/');
 					}
+					if (srcpath === 'src/examples/gist.html') {
+						return content
+							.replace(/(?:\.\.\/){3}node_modules\//g, `../node_modules/`)
+							.replace(`'../../..'`, `'..'`);
+					}
 					return content;
 				}
 			}
